backend-flask/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5
SQLAlchemy==2.0.36
pyodbc==5.3.0